feat(sdk): update go sdk path for config
