+++ Vorpal.go.toml
@@ -1,7 +1,7 @@
 language = "go"
 
 [build]
-directory = "sdk/go"
+directory = "sdk/go/cmd/vorpal"
 
 [source]
 includes = ["sdk/go"]
